backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pymongo==4.12.1
clickhouse-connect==0.8.14
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
gunicorn==23.0.0
